# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Author's Haven [https://metis-ah-staging.herokuapp.com/](https://metis-ah-staging.herokuapp.com/) is a platform that allows you to share your knowledge across the world",
    "version": "1.0.0",
    "title": "Author's Haven",
    "contact": {
      "email": "metis.ah2018@gmail.com"
    }
  },
  "host": "metis-ah-staging.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/v1/articles/{articleId}": {
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Route for authenticated users to post comments on an article",
        "description": "Returns the user and the user's comment",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "comment object to be returned when a user makes a post request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Comment posted successfully"
          },
          "400": {
            "description": "Comment is empty"
          },
          "401": {
            "description": "User access required"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [ "Notifications" ],
        "summary": "Allows user to view notifications",
        "description": "This gives user the ability to view notifications",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "authentication access token for users to access this endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [ "Notifications" ],
        "summary": "Allows user to mark all notifications as read",
        "description": "This gives user the ability to mark all his notifications as read",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "authentication access token for users to access this endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [ "Notifications" ],
        "summary": "Allows user to delete notifications hitory",
        "description": "This gives user the ability to delete all his notifications",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "authentication access token for users to access this endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Comment": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "Great post!"
        }
      },
      "json": {
        "name": "content"
      }
    },
...
  },
...
}